dom/presentation/tests/mochitest/test_presentation_device_available_change.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for the Presentation API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <!--script type="text/javascript" src="manifest.js"></script-->
</head>
<body>
<script class="testbody" type="text/javascript">

SimpleTest.expectAssertions(0, 5);

function testEnd() {
  SimpleTest.finish();
}

function testAvailability() {
  var presentation = navigator.presentation;
  var available = presentation.available;
	ok(!available, " There should be no device ! ");

  var URL = "http://example.org/presentation.html";
  var SessionId = "testLastSessionId1234";
  presentation.startSession(URL, SessionId).then(
		function onResolved(newSession) {
      ok(newSession.id == SessionId, "SessionId should be testLastSessionId1234");
			ok(true, "StartSession with last SessionId resolved.");
//      testEnd();
		},
    function onRejected() {
      ok(false, "StartSession with last SessionId rejected.");
      testEnd();
    }
  );
	presentation.onavailablechange = function(e) {
    ok(presentation.available,
     " Presentation.available should be same as e.available ");
    testEnd();
  }
}

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({"set": [["dom.presentation.enabled", true],
                                  ["dom.ignore_webidl_scope_checks", true],
                                  ["dom.presentation.test.enabled", true],
                                  ["dom.presentation.test.stage", 0]]},
  testAvailability
);
</script>
</pre>
</body>
</html>
